% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JsBarcode.R
\name{barcode_formats}
\alias{barcode_formats}
\title{Barcode formats supported by the bundled JsBarcode library}
\format{
A character vector.
}
\usage{
barcode_formats
}
\description{
Character vector with every symbology accepted by the \code{format} argument of
\code{\link[=JsBarcode]{JsBarcode()}}.
}
